Refactoring of fetch api
