+++ frontend/app/pages/login/LoginPage.tsx
@@ -1,8 +1,6 @@
 import {
   Alert,
   Button,
-  Checkbox,
-  Container,
   Paper,
   PasswordInput,
   TextInput,
